chore: Better naming of modules, project and folders
