+++ test/openai_test.exs
@@ -1,4 +1,4 @@
-defmodule OpenaiTest do
+defmodule OpenAiTest do
   use ExUnit.Case
   doctest OpenAI
